ml/requirements.txt
# ML Server Dependencies

# Core ML
ultralytics>=8.1.0
torch>=2.1.0
torchvision>=0.16.0

# gRPC
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0

# HTTP client (for CVAT, catalog sync)
requests>=2.31.0

# Configuration
pyyaml>=6.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
ruff>=0.1.0
mypy>=1.8.0
